web-demo/requirements.txt
pandas~=1.1.1
pytz~=2020.1
redis~=3.5.3
rq~=1.7.0
paramiko~=2.7.2
sshtunnel~=0.3.1
flask_excel~=0.0.7
requests~=2.24.0
bs4~=0.0.1
beautifulsoup4~=4.9.3
python-docx~=0.8.10
flask